+++ R/crmPack-package.R
@@ -33,7 +33,7 @@
 #' @importFrom graphics plot hist legend lines matlines matplot
 #' @importFrom stats binomial coef cov2cor gaussian glm lm median model.matrix
 #' optim pgamma plogis pnorm qgamma qlogis qnorm quantile rbinom rgamma approxfun
-#' rnorm runif uniroot var vcov step
+#' rnorm runif uniroot var vcov step mad
 #' @importFrom utils data head tail capture.output
 #' @importFrom lifecycle badge
 #' @importFrom rjags jags.model jags.samples
@@ -125,6 +125,7 @@ globalVariables(c("log.betaZ",
                   "tend",
                   "t0_case",
                   "tend_case",
-                  "yhat"))
+                  "yhat",
+                  "ref_dose"))
 
 # nolint end