+++ tests/testthat/test-Rules-class.R
@@ -10,4 +10,18 @@ test_that(".IncrementsNumDoseLevels works as expected", {
 test_that("IncrementsNumDoseLevels object can be created with user constructor", {
   result <- expect_silent(IncrementsNumDoseLevels())
   expect_valid(result, "IncrementsNumDoseLevels")
+
+  
+# StoppingMTDCV-class ----
+
+test_that(".StoppingMTDCV works as expected", {
+  result <- expect_silent(.StoppingMTDCV())
+  expect_valid(result, "StoppingMTDCV")
+})
+
+# StoppingMTDCV-constructor ----
+
+test_that("StoppingMTDCV object can be created with user constructor", {
+  result <- expect_silent(StoppingMTDCV())
+  expect_valid(result, "StoppingMTDCV")
 })